refactor: map 패키지 분리
