+++ android/app/src/main/AndroidManifest.xml
@@ -23,7 +23,7 @@
             android:name=".presentation.SearchActivity"
             android:exported="false" />
         <activity
-            android:name=".presentation.view.courses.CoursesActivity"
+            android:name=".presentation.view.course.CoursesActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
