+++ android/app/src/main/java/io/coursepick/coursepick/presentation/view/course/CourseViewHolder.kt
@@ -1,4 +1,4 @@
-package io.coursepick.coursepick.presentation.view.courses
+package io.coursepick.coursepick.presentation.view.course
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
