+++ android/app/src/main/java/io/coursepick/coursepick/presentation/view/map/kakao/KakaoMapLifecycleHandler.kt
@@ -1,4 +1,4 @@
-package io.coursepick.coursepick.presentation
+package io.coursepick.coursepick.presentation.view.map.kakao
 
 import com.kakao.vectormap.KakaoMap
 import com.kakao.vectormap.KakaoMapReadyCallback
