+++ android/app/src/main/java/io/coursepick/coursepick/presentation/view/map/kakao/RouteLineOptionsFactory.kt
@@ -1,4 +1,4 @@
-package io.coursepick.coursepick.presentation
+package io.coursepick.coursepick.presentation.view.map.kakao
 
 import android.content.Context
 import com.kakao.vectormap.LatLng
